backend_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
